requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt==3.2.0
pytest
httpx
pydantic
gunicorn
python-dotenv
psycopg2-binary
alembic
